Removed the confusing actions component. Tests are not fully passing right now but they need to be modified anyway
